signal_bot_cloud/requirements.txt
python-telegram-bot==20.7
apscheduler==3.10.4
aiohttp==3.9.3
numpy==1.26.4
pytz==2024.1
